Task 7.2 Updated produceInputs
1. I got rid of mpz stuff in this function there were three var/array replacements:
a. Element -> changed it to an int variable
b. Shares -> changed it to an int array
c. Elements that will be sent to convertFloat() -> changed it to an array of int
2. The functions that I got rid of and used regular computation instead are:
a. Mpz_init -> I replaced this with regular initialization
b. Mpz_set_str -> I replaced this with std::stoi to convert str to int
c. Mpz_get_str -> I used std::to_string